requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
boto3==1.35.97
botocore==1.35.97
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorlog==6.9.0
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
elevenlabs==1.50.3
exceptiongroup==1.2.2
executing==2.1.0
ffmpeg==1.4
filelock==3.16.1
fsspec==2024.12.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
moviepy==2.1.1
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.59.7
openai-whisper==20240930
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.3.1
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.0
srt==3.5.3
stack_data==0.6.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.1
wheel==0.45.1
yt-dlp==2024.12.23
zipp==3.21.0
